DISCLAIMER!!!
I really wish my name was not affiliated with this component.  From the start I believed (and voiced my opinion about) several things:
1) The design is poor and unclear.
2) The implementation is even worse.
3) Functional tests did not even compile!

I had to fix the submission and test cases in several places to get the submission to pass the tests.  I _merely_ got the submission to pass.  I still believe that the component is almost unusable in it's current state.
Were it up to me I never would have passed the submission in the first place.

Okay, I feel better with that out of my conscience.